requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
weaviate-client==3.24.1
python-dotenv==1.0.0
embed-anything==0.5.2
pydantic==2.4.2
typing-extensions==4.8.0
python-multipart==0.0.6
pymupdf4llm==0.1.2
langchain==0.0.350
gunicorn==21.2.0 